requirements.txt
flask
numpy
pyserial